Fixing some minor issues related to GCC

+++ examples/ledger/peer-to-peer-v2/main.cpp
@@ -19,7 +19,7 @@ class FetchLedger
 public:
   FetchLedger(uint16_t offset, std::string const &name, std::size_t const &shards ) :
     thread_manager_( new fetch::network::ThreadManager(64) ),
-    controller_( 1337 + offset, 7070 + offset, name, thread_manager_ )
+    controller_( uint16_t(1337 + offset), uint16_t(7070 + offset), name, thread_manager_ )
   {
     for(std::size_t i=0; i < shards; ++i)
     {
