+++ tests/math/test_exp.cpp
@@ -33,7 +33,7 @@ double test_timing(double x_value) {
   }
   auto t2a = std::chrono::high_resolution_clock::now();
 
-  double us1 = std::chrono::duration_cast<std::chrono::milliseconds>(t2a - t1a).count();
+  double us1 = double(std::chrono::duration_cast<std::chrono::milliseconds>(t2a - t1a).count());
 
 
   auto t1b = std::chrono::high_resolution_clock::now();
@@ -43,7 +43,7 @@ double test_timing(double x_value) {
   }
   auto t2b = std::chrono::high_resolution_clock::now();
 
-  double us2 = std::chrono::duration_cast<std::chrono::milliseconds>(t2b - t1b).count();
+  double us2 = double(std::chrono::duration_cast<std::chrono::milliseconds>(t2b - t1b).count());
 
   auto t1c = std::chrono::high_resolution_clock::now();
   for(std::size_t i=0; i < MAX; ++i) {
@@ -52,7 +52,7 @@ double test_timing(double x_value) {
   }
   auto t2c = std::chrono::high_resolution_clock::now();
   
-  double us3 = std::chrono::duration_cast<std::chrono::milliseconds>(t2c - t1c).count();
+  double us3 = double(std::chrono::duration_cast<std::chrono::milliseconds>(t2c - t1c).count());
   
   return (us3-us1)/(us2-us1);
   //  auto t3 = std::chrono::high_resolution_clock::now();
