+++ tests/random/bit_statistics.hpp
@@ -13,7 +13,7 @@ public:
   void operator() () {
     uint64_t s = generator_();
     for(std::size_t i=0; i < E_BIT_COUNT; ++i)
-      stats_[ i ] += (s >> i) & 1;
+      stats_[ i ] += uint32_t( (s >> i) & 1 );
     ++counter_;
   }
